+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Yayınlanan sürüm numarası yalnızca [`package.json`](package.json) `version`
 
 ## [Unreleased]
 
+### Fixed
+
+- Aylık borç vadeleri: bitiş ayındaki taksitler `range.to` ay ortasında kalsa bile grafikte ve listede gösterilir.
+
 ## [0.1.53]
 
 ### Added